requirements.txt
addict==2.4.0
albucore==0.0.13
albumentations==1.4.10
altair==5.5.0
annotated-types==0.7.0
anyconfig==0.14.0
anyio==4.6.2.post1
astor==0.8.1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
cython==0.29.37
decorator==5.1.1
editdistance==0.8.1
exceptiongroup==1.2.2
flask==3.1.0
fonttools==4.55.0
gevent==24.11.1
gevent-websocket==0.10.1
gitdb==4.0.11
gitpython==3.1.43
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
imageio==2.36.0
imgaug==0.4.0
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lazy-loader==0.4
lmdb==1.5.1
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
munch==4.0.0
narwhals==1.15.0
natsort==8.4.0
networkx==3.2.1
ninja==1.11.1.2
numpy==1.24.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.9.4.19
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.2
paddlepaddle-gpu==3.0.0b2
pandas==2.2.3
pillow==11.0.0
protobuf==5.28.3
psutil==6.1.0
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyclipper==1.3.0.post6
pydantic==2.10.1
pydantic-core==2.27.1
pydeck==0.9.1
pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytorch-cuda==11.8
pytz==2024.2
pyyaml==5.4.1
rapidfuzz==3.10.1
